+++ app/Http/Requests/Auth/ResendEmailVerificationRequest.php
@@ -5,7 +5,7 @@ namespace App\Http\Requests\Auth;
 use App\Rules\PassesRecaptcha;
 use Illuminate\Foundation\Http\FormRequest;
 
-class ForgotPasswordRequest extends FormRequest
+class ResendEmailVerificationRequest extends FormRequest
 {
     /**
      * Get the validation rules that apply to the request.
@@ -15,11 +15,6 @@ class ForgotPasswordRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'email' => [
-                'required',
-                'email',
-                'exists:users,email',
-            ],
             'recaptcha_action' => [
                 'required',
                 'string',
@@ -36,9 +31,6 @@ class ForgotPasswordRequest extends FormRequest
     {
 
         return [
-            'email.required' => 'Email is required.',
-            'email.email' => 'Email is invalid.',
-            'email.exists' => 'Email not found.',
             'recaptcha_action.required' => 'Recaptcha action is required.',
             'recaptcha_action.string' => 'Recaptcha action is invalid.',
             'recaptcha_response.required' => 'Recaptcha response is required.',