+++ app/Http/Requests/Auth/ResetPasswordRequest.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Requests\Auth;
 
+use App\Rules\PassesRecaptcha;
 use Illuminate\Foundation\Http\FormRequest;
 
 class ResetPasswordRequest extends FormRequest
@@ -34,6 +35,15 @@ class ResetPasswordRequest extends FormRequest
                 'string',
                 'in:on',
             ],
+            'recaptcha_action' => [
+                'required',
+                'string',
+            ],
+            'recaptcha_response' => [
+                'required',
+                'string',
+                new PassesRecaptcha(),
+            ],
         ];
     }
 
@@ -51,6 +61,10 @@ class ResetPasswordRequest extends FormRequest
             'password.confirmed' => 'Passwords do not match.',
             'remember.in' => 'The remember checkbox must be checked or not.',
             'remember.string' => 'The remember checkbox must be checked or not.',
+            'recaptcha_action.required' => 'Recaptcha action is required.',
+            'recaptcha_action.string' => 'Recaptcha action is invalid.',
+            'recaptcha_response.required' => 'Recaptcha response is required.',
+            'recaptcha_response.string' => 'Recaptcha response is invalid.',
         ];
     }
 }