+++ resources/views/home.blade.php
@@ -1,1 +1,6 @@
-Welcome to {{ config('app.name') }}
+@extends('layouts.app')
+@section('content')
+
+    Welcome to {{ config('app.name') }}
+
+@stop