+++ resources/views/layouts/includes/head.blade.php
@@ -1,3 +1,4 @@
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-@vite('resources/css/app.css')
+<script src="https://www.google.com/recaptcha/api.js?render={{ config('services.recaptcha.key') }}"></script>
+@vite(['resources/css/app.css', 'resources/js/app.js'])