+++ tests/Feature/Auth/ForgotPasswordViewTest.php
@@ -33,4 +33,16 @@ class ForgotPasswordViewTest extends TestCase
         $response->assertStatus(302);
         $response->assertRedirect('/dashboard');
     }
+
+    public function test_recaptcha_data_is_present()
+    {
+        $response = $this->get('/forgot-password');
+
+        $response->assertStatus(200);
+        $response->assertSeeInOrder([
+            'recaptcha-protected-form',
+            'data-sitekey',
+            'data-action',
+        ]);
+    }
 }